Silence double promotion warning (unavoidable)

+++ include/fogml_config.h
@@ -100,12 +100,12 @@ void fogml_learning(float *time_series_data) {
     start = TIME;
     tinyml_dsp(time_series_data, vector, &my_dsp_config);
     end = TIME;
-    printf("DSP took %.2f sec\n", (float)(end-start)/1000000);
+    printf("DSP took %.2f sec\n", (double)(end-start)/1000000);
 
     start = TIME;
     tinyml_reservoir_sampling(vector, &my_rs_config);
     end = TIME;
-    printf("RES took %.2f sec\n", (float)(end-start)/1000000);
+    printf("RES took %.2f sec\n", (double)(end-start)/1000000);
 
 #ifdef FOGML_VERBOSE
     //tinyml_reservoir_verbose(&my_rs_config);
@@ -119,7 +119,7 @@ void fogml_learning(float *time_series_data) {
     start = TIME;
     tinyml_lof_learn(&my_lof_config);
     end = TIME;
-    printf("LOF took %.2f sec\n", (float)(end-start)/1000000);
+    printf("LOF took %.2f sec\n", (double)(end-start)/1000000);
     
     free(vector);
 }
@@ -133,7 +133,7 @@ void fogml_processing(float *time_series_data, float *score) {
     start = TIME;
     tinyml_dsp(time_series_data, vector, &my_dsp_config);
     end = TIME;
-    printf("DSP took %.2f sec\n", (float)(end-start)/1000000);
+    printf("DSP took %.2f sec\n", (double)(end-start)/1000000);
 
 #ifdef FOGML_VERBOSE
     for(int i = 0; i < FOGML_VECTOR_SIZE; i++) {
@@ -146,7 +146,7 @@ void fogml_processing(float *time_series_data, float *score) {
     start = TIME;
     *score = tinyml_lof_score(vector, &my_lof_config);
     end = TIME;
-    printf("LOF took %.2f sec\n", (float)(end-start)/1000000);
+    printf("LOF took %.2f sec\n", (double)(end-start)/1000000);
 
 #ifdef FOGML_VERBOSE
     fogml_printf("LOF Score = ");
@@ -165,12 +165,12 @@ void fogml_classification(float *time_series_data, int* cl) {
     start = TIME;
     tinyml_dsp(time_series_data, vector, &my_dsp_config);
     end = TIME;
-    printf("RF DSP took %.2f sec\n", (float)(end-start)/1000000);
+    printf("RF DSP took %.2f sec\n", (double)(end-start)/1000000);
     
     start = TIME;
     *cl = classifier(vector);
     end = TIME;
-    printf("RF cls took %.2f sec\n", (float)(end-start)/1000000);
+    printf("RF cls took %.2f sec\n", (double)(end-start)/1000000);
 
 
 #ifdef FOGML_VERBOSE
